+++ libraries/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory( rui )
 add_subdirectory( fc )
 add_subdirectory( db )
 add_subdirectory( deterministic_openssl_rand )
